Merge upstream/main (aac4aed489d1) — header trim, jsCast removal, JSScriptFetcher removal

+++ Source/WebCore/html/track/VTTRegion.cpp
@@ -199,7 +199,7 @@ void VTTRegion::parseSettingValue(RegionSetting setting, VTTScanner& input)
         break;
     }
     case Width: {
-        float floatWidth;
+        double floatWidth;
         if (WebVTTParser::parseFloatPercentageValue(input, floatWidth) && parsedEntireRun(input, valueRun))
             m_width = floatWidth;
         else
